requirements.txt
python-dotenv
Flask==2.2.2
SQLAlchemy==2.0.36
PyMySQL==1.1.1
gunicorn==23.0.0
cloud-sql-python-connector==1.16.0
functions-framework==3.8.2
Werkzeug==2.3.8
flask-cors
